Fix conversion errors

+++ Sources/iOS/TabBar.swift
@@ -395,7 +395,7 @@ fileprivate extension TabBar {
         
         line.animate(.duration(0.25),
                      .size(CGSize(width: tabItem.width, height: lineHeight)),
-                     .position(CGPoint(x: tabItem.center.x, y: .bottom == lineAlignment ? height - lineHeight / 2 : lineHeight / 2)),
+                     .position(CGPoint(x: tabItem.center.x, y: .bottom == lineAlignment ? height - lineHeight / CGFloat(2) : lineHeight / CGFloat(2))),
                      .completion { [weak self, isTriggeredByUserInteraction = isTriggeredByUserInteraction, tabItem = tabItem, completion = completion] _ in
                         guard let s = self else {
                             return
